Messages and .aws dir

+++ provision.sh
@@ -20,10 +20,10 @@ sudo apt-get install -y libappindicator1 fonts-liberation
 wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
 sudo dpkg -i google-chrome*.deb
 
-if [ ! -f /home/vagrant/bin/aws ]; then
-	su -c "pip install awscli --upgrade --user" vagrant
+if [ ! -f /home/vagrant/bin/aws ]; then su -c "pip install awscli --upgrade --user" vagrant
 	[ -L /home/vagrant/bin/aws ] || ln -s /home/vagrant/.local/bin/aws /home/vagrant/bin/aws
 fi
+
 sudo locale-gen en_GB.UTF-8
 
 ###################### Sort out file permissions
@@ -32,8 +32,13 @@ for i in /home/vagrant/.hostssh/*; do
 	test -f /home/vagrant/.ssh/$(basename $i) || cat $i >> /home/vagrant/.ssh/$(basename $i)
 done
 
+echo "... Copied keys"
+
 chmod 600 /home/vagrant/.ssh/*
 chmod 700 /home/vagrant/.ssh
+
+echo ".. key perms"
+
 chown -R vagrant:vagrant /home/vagrant/.ssh
 chown -R vagrant:vagrant /home/vagrant/.aws
 chown -R vagrant:vagrant /home/vagrant/.gcp
@@ -42,6 +47,8 @@ find . -maxdepth 2 -type d -exec chown -R vagrant:vagrant {} \; -name ".*"
 mkdir -p /home/vagrant/tmp && chown -R vagrant:vagrant /home/vagrant/tmp
 mkdir -p /home/vagrant/bin && chown -R vagrant:vagrant /home/vagrant/bin
 touch /home/vagrant/.z && chown vagrant:vagrant /home/vagrant/.z
+
+echo "perms"
 
 ###################### Vagrant user
 chsh -s /bin/zsh vagrant
